added users API, separated APIs into separate files
