+++ API/functions/package.json
@@ -13,9 +13,12 @@
   },
   "main": "index.js",
   "dependencies": {
+    "firebase": "^9.8.1",
     "firebase-admin": "^8.10.0",
     "firebase-functions": "^3.21.0",
+    "firebase-tools": "^7.8.1",
     "react-scripts": "^3.4.1",
+    "semver": "^7.3.7",
     "uuid": "^7.0.3",
     "uuid-v4": "^0.1.0",
     "uuidv4": "^6.2.13"
