+++ API/functions/users.js
@@ -1,0 +1,42 @@
+//import { getStorage, ref, uploadBytes } from "firebase/storage";
+
+const admin = require('firebase-admin');
+const FieldValue = require('firebase-admin').firestore.FieldValue;
+const Timestamp = require('firebase-admin').firestore.Timestamp;
+const functions = require('firebase-functions');
+const express = require('express');
+const {v4: uuid} = require('uuid');
+
+
+
+// Get a reference to the storage service, which is used to create references in storage bucket
+const storage = admin.storage();
+// Create a storage reference from our storage service
+const bucket = storage.bucket();
+
+const firestore = admin.firestore();
+
+const PROJECTID = 'sociable-messenger';
+const USERS = firestore.collection('test_users');
+const STATUS = firestore.collection('status');
+const KEY = functions.config().messaging.key;
+const users_api = express();
+
+
+async function getUsers() {
+    const snapshot = await USERS.get();
+    return snapshot.docs.map(doc => doc.data());
+    //.map(doc => doc.id);
+  }
+
+users_api.get('/getUsers', async (request, response) => {
+  const req_key = request.get('auth');
+  if (req_key == KEY) {
+    await response.json(await getUsers());
+  } else {
+    response.status(401).send('Unauthorized');
+  }
+});
+
+
+exports.users_api = functions.https.onRequest(users_api)
